API/requirements.txt
flask
flask-cors
grpcio
grpcio-tools
